Wait longer before dying
This covers a case where running xdg-open doesn't launch a web browser
immediately, but an interim program to request the resource, check its
mimetype, then decice that, yes, a web browser should be launched; in
practice this means that the important request to wait for before dying
is the second one and not the first one.

In practice, waiting a few seconds more works just as fine.